requirements.txt
azure.identity
azure.mgmt.securityinsight
pydantic
PyGithub
flask
PyYAML